realsense: Move realsense cameras to default binary file
Include Realsense stream cameras in default CSD build. If system doesn't
have all required librealsense files, realsense support will be
disabled.
