+++ .gitignore
@@ -34,4 +34,3 @@ tags
 camera-sample-client
 camera-sample-custom
 camera-sample-mavlink-client
-camera-sample-realsense
