+++ Makefile.am
@@ -126,7 +126,17 @@ AM_LDFLAGS = \
 	-Wl,--no-undefined \
 	-Wl,--gc-sections
 
+REALSENSE_SRC =
+if HAVE_REALSENSE
+REALSENSE_SRC += \
+	src/stream_builder_realsense.h \
+	src/stream_builder_realsense.cpp \
+	src/stream_realsense.cpp \
+	src/stream_realsense.h
+endif
+
 BASE_FILES = \
+	$(REALSENSE_SRC) \
 	src/avahi_publisher.cpp \
 	src/avahi_publisher.h \
 	src/conf_file.cpp \
@@ -182,12 +192,7 @@ CLEANFILES += csd.service
 #Samples
 EXTRA_PROGRAMS = samples/camera-sample-custom samples/camera-sample-client samples/camera-sample-mavlink-client
 
-if HAVE_REALSENSE
-EXTRA_PROGRAMS += samples/camera-sample-realsense
-endif
-
 samples_camera_sample_custom_LDADD = ${csd_LDADD}
-samples_camera_sample_realsense_LDADD = ${csd_LDADD}
 samples_camera_sample_client_LDADD = $(GLIB_LIBS) $(AVAHI_LIBS)
 samples_camera_sample_mavlink_client_LDADD = $(GLIB_LIBS) $(AVAHI_LIBS)
 
@@ -198,14 +203,6 @@ samples_camera_sample_custom_SOURCES = \
 	samples/stream_builder_custom.cpp \
 	samples/stream_custom.cpp \
 	samples/stream_custom.h
-
-samples_camera_sample_realsense_SOURCES = \
-	$(BASE_FILES) \
-	src/main.cpp \
-	samples/stream_builder_realsense.h \
-	samples/stream_builder_realsense.cpp \
-	samples/stream_realsense.cpp \
-	samples/stream_realsense.h
 
 samples_camera_sample_client_SOURCES = \
 	samples/main_sample_client.cpp \
